+++ simulation.h
@@ -11,14 +11,14 @@ int randrange(int min, int max);
 // a definition of the variables and a declaration of the functions that relates the simulation
 namespace Simulation
 {
-    const int RADIUS_OF_MEDIUM = 5;  // the radius of the particles that forms the medium(e.g. a H2O molecule)
-    const int RADIUS_OF_PARTICLES = 50;  // the radius of the particles that floats in the medium(e.g. pollen grain)
-    const int MASS_OF_MEDIUM = 1;  // the mass of the particle that forms the medium(e.g. a H2O molecule)
-    const int MASS_OF_PARTICLES = 10;  // the mass of the particle that floats in the medium(e.g. pollen grain)
+    const double RADIUS_OF_MEDIUM = 5;  // the radius of the particles that forms the medium(e.g. a H2O molecule)
+    const double RADIUS_OF_PARTICLES = 50;  // the radius of the particles that floats in the medium(e.g. pollen grain)
+    const double MASS_OF_MEDIUM = 1;  // the mass of the particle that forms the medium(e.g. a H2O molecule)
+    const double MASS_OF_PARTICLES = 10;  // the mass of the particle that floats in the medium(e.g. pollen grain)
     const int NUMBER_OF_MEDIUM = 1;  // the amount of particles that forms the medium(e.g. a H2O molecule)
     const int NUMBER_OF_PARTICLES = 0;  // the number of the particles that floats in the medium(e.g. pollen grain)
-    const int MAX_SPEED = 5;  // a maximal possible speed of the particle after the start of the simulation
-    const int MIN_SPEED = 1;  // a minimal possible speed of the particle after the start of the simulation
+    const double MAX_SPEED = 5;  // a maximal possible speed of the particle after the start of the simulation
+    const double MIN_SPEED = 1;  // a minimal possible speed of the particle after the start of the simulation
     const int SURFACE_WIDTH = 640;  // width of the surface for the simulation
     const int SURFACE_HEIGHT = 480;  // height of the surface for the simulation
